fix: separating fetching and hook
